Add Task to Eff - Also move Eff to a shared package object
